Rename functions into modes
