+++ web/lib/modes/basic.js
